Comencem

+++ ex00/megaphone.cpp
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   megaphone.cpp                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ojimenez <ojimenez@student.42barcel>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/02/13 12:02:32 by ojimenez          #+#    #+#             */
+/*   Updated: 2024/02/13 12:02:36 by ojimenez         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include <iostream> //Entrada y salida standard
 #include <string> //Manejar cadenas de caracteres std::string
 #include <locale> //Configuraciones regionales
@@ -6,7 +18,7 @@ int	main(int argc, char **argv)
 {
 	std::string	s;
 	int			i = 1;
-	int			j = 0;
+	size_t		j = 0;
 
 	if (argc == 1)
 		std::cout << "* LOUD AND UNBEARABLE FEEDBACK NOISE *";
